Refactor MemorizeDocument structure and update WordBlur component to use new text item interface
