+++ app/src/main/java/net/bible/android/control/page/ClientPageObjects.kt
@@ -176,11 +176,12 @@ class MultiFragmentDocument(private val osisFragments: List<OsisFragment>, priva
         )
 }
 
-class MemorizeDocument(private val key: String, private val texts: List<Pair<String, String>>): Document {
+class MemorizeDocument(private val title: String, private val texts: List<Pair<String, String>>): Document {
     override val asHashMap: Map<String, Any>
         get() = mapOf(
             "id" to wrapString(randomUUID().toString()),
             "type" to wrapString("memorize"),
+            "title" to wrapString(title),
             "texts" to listToJson(texts.map { "{ 'key': " + wrapString(it.first) + ", 'text':" + wrapString(it.second) + "}" })
         )
 }
